requirements.txt
# Core dependencies
selenium==4.16.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
pyyaml==6.0.1

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Utility
python-dotenv==1.0.0
tqdm==4.66.1
tenacity==8.2.3
lxml==4.9.3

# Testing and linting
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0 